--- src/components/userList/store/userListStore.js
@@ -1,9 +1,0 @@
-'use strict';
-
-class UserListStore {
-  constructor() {
-    this.userList = [];
-  }
-}
-
-export const userListrSrore = new UserListStore();
